+++ packages/expo-video/android/src/main/java/expo/modules/video/VideoView.kt
@@ -14,6 +14,9 @@ import android.view.ViewGroup
 import android.widget.FrameLayout
 import android.widget.ImageButton
 import androidx.fragment.app.FragmentActivity
+import androidx.media3.common.Format
+import androidx.media3.common.MimeTypes
+import androidx.media3.common.Tracks
 import androidx.media3.ui.PlayerView
 import com.facebook.react.common.annotations.UnstableReactNativeAPI
 import com.facebook.react.modules.i18nmanager.I18nUtil
@@ -27,12 +30,13 @@ import expo.modules.kotlin.views.ExpoView
 import expo.modules.video.drawing.OutlineProvider
 import expo.modules.video.enums.ContentFit
 import expo.modules.video.player.VideoPlayer
+import expo.modules.video.player.VideoPlayerListener
 import expo.modules.video.utils.ifYogaDefinedUse
 import java.util.UUID
 
 // https://developer.android.com/guide/topics/media/media3/getting-started/migration-guide#improvements_in_media3
 @androidx.annotation.OptIn(androidx.media3.common.util.UnstableApi::class)
-class VideoView(context: Context, appContext: AppContext) : ExpoView(context, appContext) {
+class VideoView(context: Context, appContext: AppContext) : ExpoView(context, appContext), VideoPlayerListener {
   val id: String = UUID.randomUUID().toString()
   val playerView: PlayerView = PlayerView(context.applicationContext)
   val onPictureInPictureStart by EventDispatcher<Unit>()
@@ -42,6 +46,8 @@ class VideoView(context: Context, appContext: AppContext) : ExpoView(context, ap
 
   var willEnterPiP: Boolean = false
   var isInFullscreen: Boolean = false
+    private set
+  var showsSubtitlesButton = false
     private set
 
   private val currentActivity = appContext.throwingActivity
@@ -96,13 +102,15 @@ class VideoView(context: Context, appContext: AppContext) : ExpoView(context, ap
     }
 
   var videoPlayer: VideoPlayer? = null
-    set(videoPlayer) {
+    set(newPlayer) {
       field?.let {
         VideoManager.onVideoPlayerDetachedFromView(it, this)
       }
-      playerView.player = videoPlayer?.player
-      field = videoPlayer
-      videoPlayer?.let {
+      videoPlayer?.removeListener(this)
+      newPlayer?.addListener(this)
+      playerView.player = newPlayer?.player
+      field = newPlayer
+      newPlayer?.let {
         VideoManager.onVideoPlayerAttachedToView(it, this)
       }
     }
@@ -110,6 +118,7 @@ class VideoView(context: Context, appContext: AppContext) : ExpoView(context, ap
   var useNativeControls: Boolean = true
     set(value) {
       playerView.useController = value
+      playerView.setShowSubtitleButton(value)
       field = value
     }
 
@@ -236,6 +245,25 @@ class VideoView(context: Context, appContext: AppContext) : ExpoView(context, ap
     }
     rootViewChildrenOriginalVisibility.clear()
     this.addView(playerView)
+  }
+
+  override fun onTracksChanged(player: VideoPlayer, tracks: Tracks) {
+    showsSubtitlesButton = hasSubtitles(tracks)
+    playerView.setShowSubtitleButton(showsSubtitlesButton)
+    super.onTracksChanged(player, tracks)
+  }
+
+  private fun hasSubtitles(tracks: Tracks): Boolean {
+    for (group in tracks.groups) {
+      for (i in 0..<group.length) {
+        val format: Format = group.getTrackFormat(i)
+
+        if (MimeTypes.isText(format.sampleMimeType)) {
+          return true
+        }
+      }
+    }
+    return false
   }
 
   private fun calculateRectHint() {